+++ src/main/webapp/js/userProfile.js
@@ -135,25 +135,27 @@ var program = {
      * 删除用户
      */
     deleteUser: function (vals) {
-        $.ajax({
-            type: "get",
-            content: "application/x-www-form-urlencoded;charset=UTF-8",
-            url: "../userProfile/deleteByIds",
-            dataType: 'json',
-            async: false,
-            data: {
-                ids: vals
-            },
-            success: function (result) {
-                console.log(result);
-                if (result.status > 0) {
-                    pubMeth.alertInfo("alert-success", "删除成功！");
-                    program.selectProfile();
-                } else {
-                    pubMeth.alertInfo("alert-danger", "删除失败！");
+        if (confirm("你确定要删除这些" + vals + "学生吗？")) {
+            $.ajax({
+                type: "get",
+                content: "application/x-www-form-urlencoded;charset=UTF-8",
+                url: "../userProfile/deleteByIds",
+                dataType: 'json',
+                async: false,
+                data: {
+                    ids: vals
+                },
+                success: function (result) {
+                    console.log(result);
+                    if (result.status > 0) {
+                        pubMeth.alertInfo("alert-success", "删除成功！");
+                        program.selectProfile();
+                    } else {
+                        pubMeth.alertInfo("alert-danger", "删除失败！");
+                    }
                 }
-            }
-        });
+            });
+        }
     },
     /**
      * 更新信息
